% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_submission_package.R
\name{add_comments_glatos_workbook_xlsx}
\alias{add_comments_glatos_workbook_xlsx}
\title{Add comments to a glatos_workbook in openxlsx Workbook format}
\usage{
add_comments_glatos_workbook_xlsx(wbx)
}
\arguments{
\item{wbx}{An \code{\link[openxlsx::openxlsx]{openxlsx}} \code{Workbook}
object; created by \code{\link{as_glatos_workbook_xlsx}} or internal
function \code{new_glatos_workbook_xlsx}.}
}
\value{
An \code{\link[openxlsx::openxlsx]{openxlsx}} \code{Workbook} object.
}
\description{
Internal function called by \code{\link{as_glatos_workbook_xlsx}}.
}
\details{
Content for comments are read from GLATOS Data Dictionary:
\code{data\{GLATOSWeb_data_dictionary\}}.
}
\author{
Christopher Holbrook, \email{cholbrook@usgs.gov}
}
