% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_submission_package.R
\name{as_glatos_workbook_xlsx}
\alias{as_glatos_workbook_xlsx}
\title{Convert glatos_workbook object to openxlsx Workbook}
\usage{
as_glatos_workbook_xlsx(wb, local_tzone)
}
\arguments{
\item{wb}{A \code{glatos_workbook} object; e.g., created by
\code{\link{read_glatos_workbook}}.}

\item{local_tzone}{A character string specifying the time zone to be used for
the conversion from UTC to local time (e.g, "US/Eastern"). See details.}
}
\value{
An \code{\link[openxlsx::openxlsx]{openxlsx}} \code{Workbook} object.
}
\description{
Used for writing a GLATOS workbook submission package to disk.
}
\details{
Local time zone must be specified (via \code{local_tzone}) because
\code{glatos_workbook} objects only contain time stamps in UTC. However,
the GLATOS Data Submission Package (a *.zip archive containing a *.xlsm
file) requires data in a local time zone (limited to \code{"US/Eastern"} or
\code{"US/Central"}, which are actually expressed simply as
\code{"Eastern"} or \code{"Central"} in a GLATOS Submission Package. Thus,
for GLATOS Data Portal compatibility, only values of \code{local_tzone}
supported are \code{"US/Eastern"} or \code{"US/Central"} are allowed.

\code{as_glatos_workbook_xlsx} is essentially the inverse of
\code{\link[glatos::read_glatos_workbook]{read_glatos_workbook}}, resulting
in an object that resembles a GLATOS Workbook *.xlsm file in structure;
namely six worksheets: 'Project', 'Locations', 'Proposed', 'Deployment'.
Data for the 'Project' sheet are taken from \code{wb$metadata}. The
'Proposed' sheet contains no data. Data for the 'Locations', 'Deployment',
and 'Recovery' sheets are taken from \code{wb$receivers}. Data for the
'Tagging' sheet are taken from \code{wb$animals}.
}
\note{
Project-specific columns and elements of \code{wb} are not exported.
}
\examples{

#get path to example GLATOS Data Workbook
wb_file <- system.file("extdata",
                       "walleye_workbook.xlsm", package = "glatos")

wb <- glatos::read_glatos_workbook(wb_file)

wbx <- as_glatos_workbook_xlsx(wb, local_tzone = "US/Eastern")

openxlsx::openXL(wbx)

}
\author{
Christopher Holbrook, \email{cholbrook@usgs.gov}
}
