% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLATOSWeb_data_dictionary}
\alias{GLATOSWeb_data_dictionary}
\title{Data dictionary for GLATOSWeb Submission Package}
\format{
A data frame with 141 rows and 8 columns: \describe{
\item{sheet}{Worksheet name} \item{field_number}{Column number for all
sheets except Project} \item{field_name}{Column name}
\item{example}{Example value} \item{format_units}{Description of format and
units} \item{definition}{Definition} \item{GLATOS}{Required by GLATOS (Data
Portal)?} \item{OTN}{Required by Ocean Tracking Network?} }
}
\source{
\url{https://www.glatos.glos.us}
}
\usage{
GLATOSWeb_data_dictionary
}
\description{
A dataset with name, example, format, units, and network requirements
(GLATOS, OTN) of fields (columns) in each sheet of a GLATOS Data Workbook.
}
\note{
A few corrections were made to the \code{Field.#} column in the original
workbook before loading into this package.

Used to add comments to \code{openxlsx} \code{Workbook} by
\code{write_glatos_submission_package}.
}
\keyword{datasets}
