% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{GLATOSWeb_DataDictionaries}
\alias{GLATOSWeb_DataDictionaries}
\title{GLATOSWeb_DataDictionaries.xlsx}
\description{
This is a data dictionary provided by the GLATOS Data Portal.
}
\section{GLATOSWeb_DataDictionaries.xlsx}{


This data was used to create \link{GLATOSWeb_data_dictionary}
}

