% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{HECWL_20220505}
\alias{HECWL_20220505}
\title{HECWL_20220505.csv}
\description{
This is an example of file containing GPS data exported from Trimble in a
very specific peculiar format.
}
\section{HECWL_20220505.csv}{


This data is an example of the format parsed by
\link{read_henrys_trimble_csv}.
}

