% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_submission_package.R
\name{new_glatos_workbook_xlsx}
\alias{new_glatos_workbook_xlsx}
\title{Internal function to initialize openxlsx Workbook from a glatos_workbook}
\usage{
new_glatos_workbook_xlsx()
}
\value{
An \code{\link[openxlsx::openxlsx]{openxlsx}} \code{Workbook} object.
}
\description{
Called by \code{\link{as_glatos_workbook_xlsx}}.
}
\author{
Christopher Holbrook, \email{cholbrook@usgs.gov}
}
