% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-gps.R
\name{read_henrys_trimble_csv}
\alias{read_henrys_trimble_csv}
\title{Read Henry's peculiar Trimble export CSV format}
\usage{
read_henrys_trimble_csv(gps_file_in)
}
\arguments{
\item{gps_file_in}{}
}
\value{
A data.frame with site, latitude, longitude,
timestamp_utc, and source_file.
}
\description{
Read Henry's peculiar Trimble export CSV format
}
\details{
Specific to HBBS. Reads GPS data exported from a Trimble in a
format like this: \cr\cr

"Site ","MCK001","2022/05/05","09:11:26"
" ",45.794284940,-84.770085514,2022/05/05 13:11:44
" ",45.794280571,-84.770072229,2022/05/05 13:11:45
" ",45.794276437,-84.770057981,2022/05/05 13:11:46
" ",45.794272347,-84.770043991,2022/05/05 13:11:47
" ",45.794267438,-84.770029965,2022/05/05 13:11:48
"Site ","MCK002","2022/05/05","09:30:22"
" ",45.801236531,-84.763650723,2022/05/05 13:30:40
" ",45.801237444,-84.763642311,2022/05/05 13:30:41
" ",45.801238197,-84.763634249,2022/05/05 13:30:42
" ",45.801239154,-84.763625886,2022/05/05 13:30:43
" ",45.801240122,-84.763617542,2022/05/05 13:30:44
}
\examples{

# Example file
gps_file <- system.file("inst/extdata", "HECWL_20220505.csv",
                        package = "glatosExtras")

gps <- read_henrys_trimble_csv(gps_file)


\dontrun{
# maybe followed by something like this (for median location):

library(dplyr)

gps \%>\%
  group_by(site, source_file) \%>\%
  summarize(longitude = median(longitude),
            latitude = median(latitude),
            timestamp_utc = median(timestamp_utc))

}

returns
}
