% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_submission_package.R
\name{style_glatos_workbook_xlsx}
\alias{style_glatos_workbook_xlsx}
\title{Internal function to style glatos_workbook in openxlsx Workbook format}
\usage{
style_glatos_workbook_xlsx(wbx)
}
\arguments{
\item{wbx}{An \code{\link[openxlsx::openxlsx]{openxlsx}} \code{Workbook}
object; created by \code{\link{as_glatos_workbook_xlsx}} or internal
function \code{new_glatos_workbook_xlsx}.}
}
\value{
An \code{\link[openxlsx::openxlsx]{openxlsx}} \code{Workbook} object.
}
\description{
Called by \code{\link{as_glatos_workbook_xlsx}} and internal function
\code{new_glatos_workbook_xlsx}.
}
\author{
Christopher Holbrook, \email{cholbrook@usgs.gov}
}
