% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_submission_package.R
\name{write_glatos_workbook_xlsx}
\alias{write_glatos_workbook_xlsx}
\title{\link{Deprecated} Use write_glatos_workbook instead}
\usage{
write_glatos_workbook_xlsx(...)
}
\description{
This function has been deprecated and will be removed from next
version of package. Use \code{\link{write_glatos_workbook}} instead
}
