% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatos_submission_package.R
\name{xlsx_to_xlsm}
\alias{xlsx_to_xlsm}
\title{Convert XLSX to XLSM file}
\usage{
xlsx_to_xlsm(xlsx_file)
}
\description{
Convert XLSX to XLSM file
}
\note{
I am not sure there is any need for this function.
}
